Moved packages out of options/.  Tests are very broken.
