+++ internal/security/secrets/command/cache/cache_test.go
