+++ internal/security/secrets/command/generate/generate_test.go
