+++ internal/security/secrets/command/import/import.go
@@ -20,8 +20,8 @@ import (
 	"flag"
 	"fmt"
 
-	"github.com/edgexfoundry/edgex-go/internal/security/secrets/option/contract"
-	"github.com/edgexfoundry/edgex-go/internal/security/secrets/option/helper"
+	"github.com/edgexfoundry/edgex-go/internal/security/secrets/contract"
+	"github.com/edgexfoundry/edgex-go/internal/security/secrets/helper"
 
 	"github.com/edgexfoundry/go-mod-core-contracts/clients/logger"
 )
