+++ internal/security/secrets/command/legacy/flagset.go
