+++ internal/security/secrets/contract/command.go
